+++ src/test/java/racingcar/CarControllerTest.java
@@ -3,7 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.Car;
-import racingcar.domain.CarController;
+import racingcar.controller.CarController;
 import racingcar.domain.CarList;
 
 import java.util.ArrayList;
